backend/requirements.txt
# Restaurant Lab API — install: pip install -r requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pymysql>=1.1.0
cryptography>=43.0.0
alembic>=1.14.0
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
PyJWT>=2.9.0
bcrypt>=4.2.0
email-validator>=2.2.0
httpx>=0.27.0
langchain-core>=1.2.20
